% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_batts}
\alias{find_batts}
\title{Make educated guesses for which variables pertain to the same battery}
\usage{
find_batts(data)
}
\description{
Make educated guesses for which variables pertain to the same battery
}
