% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oatmeal-package.R
\docType{package}
\name{oatmeal-package}
\alias{oatmeal}
\alias{oatmeal-package}
\title{oatmeal: Utilities for Working with Survey Data}
\description{
Low-level functions for helping with commong survey data wrangling tasks
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jjmoncus.github.io/oatmeal/}
}

}
\author{
\strong{Maintainer}: J.J. Moncus \email{jjmoncus706@gmail.com}

}
\keyword{internal}
